MB-20323: newFileName to be updated only during compaction
Falling back to the old way of updating file_status
and oldFileName as newFileName for a file need not
be updated anywhere except during compaction. With
this in mind, getting rid of the updateFileLinkage
method, and accommodating updateFileStatus to update
oldFileName as done before the change that introduced
the new way of tracking compacted files.

Change-Id: I2dbbd571006aee4f3701eb50f88cd5931f53975e